Refactor Python file fetchers to use SharedFileFetcher
